project_nexus_jobboard/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
python-dotenv
drf-yasg
django-filter
django-extensions
